+++ addons/clyde/examples/simple_example/example.gd
@@ -1,6 +1,7 @@
 extends MarginContainer
 
 var _dialogue
+var _external_persistence = {}
 
 onready var _line = $PanelContainer/MarginContainer/line
 onready var _options = $PanelContainer/MarginContainer/options
@@ -12,6 +13,9 @@ func _ready():
 
 	_dialogue.connect("event_triggered", self, '_on_event_triggered')
 	_dialogue.connect("variable_changed", self, '_on_variable_changed')
+
+	_dialogue.on_external_variable_fetch(funcref(self, "_on_external_variable_fetch"))
+	_dialogue.on_external_variable_update(funcref(self, "_on_external_variable_update"))
 
 
 func _get_next_dialogue_line():
@@ -75,3 +79,16 @@ func _on_restart_pressed():
 	_dialogue_ended_message.hide()
 	_dialogue.start()
 	_get_next_dialogue_line()
+
+
+# this is an example on how to provide access to external variables.
+# The dialogue used in this example does not use external variables, but for instance,
+# if it tried to access { @health }, this method would be called and return the value from
+# _external_persistence["health"]
+func _on_external_variable_fetch(variable_name: String):
+	return _external_persistence[variable_name]
+
+
+# This method is called when the dialogue tries to set an external variable. i.e { set @health = 10 }
+func _on_external_variable_update(variable_name: String, value):
+	_external_persistence[variable_name] = value