+++ addons/clyde/interpreter/Memory.gd
@@ -1,18 +1,18 @@
 extends Reference
 
 signal variable_changed(name, value, previous_value)
-signal external_variable_changed(name, value, previous_value)
 
 const SPECIAL_VARIABLE_NAMES = [ 'OPTIONS_COUNT' ];
 
 var _mem = {
 	"access": {},
 	"variables": {},
-	"e_variables": {},
 	"internal": {}
 }
 
 var _external_variable_prefix = "@"
+var _external_variable_fetch_callback
+var _external_variable_update_callback
 
 func set_as_accessed(id):
 	_mem.access[str(id)] = true
@@ -27,7 +27,7 @@ func get_variable(id, default_value = null):
 		return get_internal_variable(id, default_value);
 
 	if id.begins_with(_external_variable_prefix):
-		return get_external_variable(id, default_value)
+		return get_external_variable(id)
 
 	var value = _mem.variables.get(id);
 	if (value == null):
@@ -58,17 +58,19 @@ func get_internal_variable(id, default_value):
 
 func set_external_variable(id: String, value):
 	var sanitized_id = id.replace(_external_variable_prefix, "")
-	var old_var = _mem.e_variables.get(sanitized_id)
-	emit_signal("external_variable_changed", sanitized_id, value, old_var)
-	_mem.e_variables[sanitized_id] = value
+	if _external_variable_update_callback is FuncRef:
+		_external_variable_update_callback.call_func(sanitized_id, value)
+
 	return value
 
 
-func get_external_variable(id: String, default_value = null):
+func get_external_variable(id: String):
 	var sanitized_id = str(id).replace(_external_variable_prefix, "")
-	var value = _mem.e_variables.get(sanitized_id)
-	if value == null:
-		return default_value;
+	var value
+
+	if _external_variable_fetch_callback is FuncRef:
+		value = _external_variable_fetch_callback.call_func(sanitized_id)
+
 	return value
 
 
@@ -82,13 +84,19 @@ func get_all():
 
 func load_data(data):
 	_mem = data
-	_mem["e_variables"] = {}
 
 
 func clear():
 	_mem = {
 		"access": {},
 		"variables": {},
-		"e_variables": {},
 		"internal": {}
 	}
+
+
+func on_external_variable_fetch(callback: FuncRef) -> void:
+	_external_variable_fetch_callback = callback
+
+
+func on_external_variable_update(callback: FuncRef) -> void:
+	_external_variable_update_callback = callback