+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## Unreleased (2024-)
+## 5.0.0 (2024-09-11)
 
 ### Breaking changes